Make DOT output of whole graphs more consistent (#783)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/783

When is the best time to make a change that causes a bunch of test output to need updating? YESTERDAY. But today is a close second.

The DOT visualizer for graphs has two modes: one where it displays only those nodes that are ancestors of a query/factor/observation/sample, and one where it displays all the nodes in the graph, even the orphans.

There were two problems here, one conceptual and one practical. The conceptual problem is that those two modes used completely different algorithms for enumerating the graph and producing the DOT output. The "just ancestors" algorithm did a topological sort where nodes "on the left" are before nodes "on the right", and ancestors are enumerated before descendants. The "whole graph" algorithm enumerated nodes in the order they were added to the graph.

This then causes a small but noticable quirk where the graph visualizer often puts "left" nodes on the right side of the graph when it draws it.

The practical problem is that the "whole graph" algorithm used an internal data structure of the graph builder which I might want to change in the future.

By changing the whole graph algorithm to also do a topo sort we eliminate all these issues. The output order becomes consistent and based on the topology of the graph, and we no longer consume an internal data structure.

The down side is that this changed the output of about twenty tests. I've verified that the output is correct and updated all the tests; hopefully this is the last time we do so.

Reviewed By: wtaha

Differential Revision: D27473148

fbshipit-source-id: e5996e8b5e202ab300d82b9ca461bf8bfc35ab2e