Split graph builder into two classes
Summary:
In the previous diff I noted that half the functions in the graph builder class are used only for building the graph, and half the functions are used as runtime hooks when the graph accumulator is running. These tasks are logically seperable into two classes, and so I have done so. BMGraphBuilder is now responsible only for creating a deduplicated graph. BMGRuntime is responsible for intercepting function calls, rewriting their source code, and intercepting the function calls in the rewritten code.

There is still some more minor refactoring to do here to make the interface clean and stop the various classes from looking at each other's internals, but this split is a large step towards making all this code more targetted at a specific concern.

In the next diff I'll move these into separate modules.

Reviewed By: wtaha

Differential Revision: D27481957

fbshipit-source-id: 74bcee8a2e4c13873aa9815b00446c4d28121b4d