Remove pass which eliminates subtractions
Summary:
Bean Machine Graph has no subtraction operator; it represents `x-y` as `x+(-y)`.

When adding the instrumentation to a function that causes it to accumulate the graph when executed, we first did a *syntactic* transformation, so we would generate `z = x - y` as

    t = handle_negate(y)
    z = handle_add(x, t)

But this has an important problem: what if x and y are tensors, and instead we have `z = x.sub(y)` ?

We cannot use a syntactic trick to eliminate subtractions from the lifted code; we will need to handle a variety of different subtraction operations that could be in the model.

I've eliminated the syntactic rewrite, and we now generate `z = handle_subtraction(x, y)` as we should. The runtime then detects usage of `Tensor.sub` and invokes `handle_subtraction` for it too.

This change deletes the last user of the constant folding code, so I'll delete that in a later diff.

Reviewed By: wtaha

Differential Revision: D27486823

fbshipit-source-id: 5590f1c148a168d14cdb7141ef0ff7f38baa63c7