Delete constant folder module and tests
Summary:
As noted in the previous diff, the only consumer of the constant folding module was test cases: one test in the single assignment tests, and the constant folding tests.

I've updated the single assignment test to no longer use the constant folder, and deleted the rest.

The original idea here was (1) validate that the rewriting library was sufficient to solve real problems, and (2) do simple optimizations at the source code level. We did validate it, so that is no longer necessary, and we now do constant folding and other simple "peephole" optimizations in the graph construction code. Doing the same syntactically is not useful, so we can delete this code.

Reviewed By: wtaha

Differential Revision: D27488212

fbshipit-source-id: 915d7bd811a3f9842827f8df320f69e979a2466f