Move simple arithmetic optimizations to optimization pass (#790)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/790

There are a number of problems with how we compute the inf type of addition, multiplication and power nodes that I have addressed in this diff.

First let me say what the goal is. The basic idea of the inf type computation is to answer the question "what is the smallest lattice type could this node be, based on its inputs?"  We need to know that so that (1) we can insert type conversion nodes as necessary to meet the type rules of BMG, and (2) we need to produce errors for the cases where we cannot.

The problems I've fixed here are:

* Some of this code was written when we supported arithmetic operations on tensors in BMG. BMG no longer has an unbounded tensor type, so this logic was wrong.

* The inf type computation should never say that *the smallest type a node can be* is a type that it cannot possibly be! But there were cases where that could happen.

* The inf type computation attempted in several cases to answer the question "if we replaced this graph node with a different node that computed the same value, what is the smallest type THAT thing could be?"  Attempting to do so is, first off, pointless; we don't need to know what type a *different* node could be, we need to know what type *this* node could be. And second, attempting to do so requires the type computation to know what optimization passes might do in the future!

The more sensible approach is to say that for the addition, multiplication and power operators:

* If the inputs' types meet the requirements of the node in BMG, then the inf type is computed based on those types, and is the smallest possible output type.

* If the inputs' types do NOT meet those requirements then just return the largest legal type. For instance, if we have a sum of a real value and a simplex matrix, we CANNOT meet the requirements of an addition, but we must not lie and say "this sum is a tensor".  We say "this sum is a real", put a requirement on the edges that they both be reals, and then when the simplex cannot be converted to a real, we report an error.

This practice of nodes always reporting a legal type as their inf types will help prevent cascading error reports in requirements checking; if an addition node cannot be legal then we want to report that ONCE; we don't want that to then trigger a report that the consumer of the sum ALSO has a type error on the output edge.

To fix these problems:

* All arithmetic optimizations on sum, product and power nodes (other than constant folding) are moved out of node construction and requirements checking; they are moved into an arithmetic optimization class.

* Computing the inf types of these nodes no longer lies; it always reports a type that the node actually could be.

Reviewed By: wtaha

Differential Revision: D27571429

fbshipit-source-id: 12788663b26c5f6321076798f382a71e31b34f3c