Decouple runtime module from type system; fix bug in control flow generation
Summary:
I wish to decouple all graph accumulation logic in both the runtime and the builder from the type system; the type system is something we impose upon an accumulated graph, not something we use while accumulating it.

The only remaining place we use the type system in the runtime is in generating an optimized map for a stochastic control flow, and all this code is broken and wrong anyways, so this is a temporary fix at best. The only information we need at this point is whether or not a constant is zero or one, which we can compute without the type system. All this code will need to be rewritten to get proper stochastic workflows going, at which point we can move the optimization out of the runtime entirely and into an optimization pass.

While looking at that code I realized that it was even more broken than I had supposed: I swapped the consequence and alternative of the if-then-else nodes! I fixed that error and updated the tests.

Reviewed By: wtaha

Differential Revision: D27606513

fbshipit-source-id: 9b1c1e462858cc7f7c090485b9d622383caec190