Move timeseries module into facebook folder (#792)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/792

In D27183047 (https://github.com/facebookincubator/beanmachine/commit/2f7f59c892329b0959af89c3c970ca710db78633), we decided to move the timeseries module to the applications folder (from the facebook folder initially) which exposed this module to Github (which is fine) but had the inadvertent effect of including this in our CircleCI tests and breaking the build. The timeseries module, as it currently stands, doesn't depend on beanmachine, so there is no requirement to run these in CircleCI.

I initially decided to ignore the timeseries module in the parent pytest-ini config so that these tests are not run in CircleCI. However, I believe, there is no strong reason why we need to expose this module in our open source repo currently so I decided to move this back into `facebook`. Reviewers: please let me know if you think otherwise. At a later time, when open sourcing Bean Machine, we can determine whether it makes sense to include this library. At that point, we can move this back into applications and enable lint/unit tests in CircleCI.

Reviewed By: jpchen, horizon-blue

Differential Revision: D27698847

fbshipit-source-id: ad6289e8bb341a58e44f4f6738826fb78a4f16d4