Bump PyTorch version in CircleCI config (#793)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/793

Currently CircleCI is broken because the version of torch nightly we pinned was not available anymore.

Now that we don't rely on any of the experimental features, let's pin PyTorch version to 1.8.1 on GitHub so that we don't need to keep worrying about experimental release break our build.

(A note on the change in `tensorops.py`: as of torch 1.8.1, `torch.vmap` is still not available so I need to import it from `torch._vmap_internals`.)

Reviewed By: feynmanliang, neerajprad

Differential Revision: D27746321

fbshipit-source-id: 25f24fa0e521079cdbd2ba53bc2855dafcb5cc62