Create a base class to implement a pluggable node typing algorithm.
Summary:
We currently have several typing questions we need to answer about nodes, but the infrastructure we have in place for that is suboptimal:

* Logic is spread out amongst many node classes, rather than isolated in one place
* Adding a new kind of typing is not easy
* We compute type information long before we need it
* It is not clear what code needs what type information
* We have very strict requirements on the algorithms which compute type information -- they must be fast, they must not be recursive, they must handle graph mutations -- but it is hard to understand the algorithm because it is so spread out.

This diff is the start of a stack of diffs which will extract computation of the "inf type" and "graph type" from the node hierarchy and into modules specifically built to handle those problems in a principled manner.

Here I have built a base class which encapsulates the graph traversals and computations needed to add a new node to a set of type judgments and update those judgments when the shape of the graph changes.  All that is missing is the actual rule which assigns types; that will be implemented by derived classes in upcoming diffs.

Reviewed By: wtaha

Differential Revision: D27628545

fbshipit-source-id: f16ddde4c925aab76fe49456a3dd226ee307d8e7