Create a typer object to replace computation of "inf type" (#796)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/796

During the "problem fixing" phase of compilation we often need to know the "inf type" of a node. The inf type answers the question "suppose we inserted conversion nodes into this graph to make it a legal BMG graph; in that graph, what would the SMALLEST lattice type of this node be?"

For example, a sum of a probability and a natural is not a legal sum node in BMG.  We could convert the addends to positive real or to real and we would have a legal sum node.  The smaller choice is positive real, so that is the "infimum type" -- the smallest possible lattice type.

This is very useful for transforming the graph into a legal BMG graph if possible, and for reporting errors if it is not.  However our existing implementation has some problems:

* We compute the inf type of nodes even if those nodes are not legal BMG nodes; basically we are *guessing* what the inf type *ought to be* if the node were actually in BMG. This is silly. We will never *generate* a graph that contains a node that is not present in the BMG node set, so why would we need to *type* such nodes?  Moreover, does it even make sense to type a graph node that *descends* from an unsupported node? Again, that graph node cannot ever be generated into a BMG graph, so why should we spend any effort computing its type?

* The terminology of "inf type", though principled, is a little highfalutin. We can be more approachable. I'm going to call it the "lattice type" from now on.

* We have many constraints on the implementation details for correctness and performance reasons: determining the lattice type of a node might require that we know the lattice types of all ancestor nodes. Computing that is potentially expensive, so it should be cached. The path from a given node to its "highest" ancestor may involve thousands of edges, so the algorithm must not be recursive. And the graph is mutable, so we must invalidate the cache when an ancestor node changes its type due to a mutation.

The existing implementation that computes the inf type is spread out across dozens of node types and the algorithm which maintains the invariants described above is difficult to follow. Fortunately, we have encapsulated that algorithm into the TyperBase base class.

This diff begins to solve all these problems.

* We now have a single class responsible for working out the lattice type associated with a node.
* The base class maintains a cache of node types. It invalidates it on request, and recomputes the minimal set of types necessary for a given change.
* We make no attempt to assign a type to any node unsupported by BMG, or any descendent of such nodes.

We do not yet *use* this class for anything.  In upcoming diffs I will gradually replace uses of `inf_type` with calls into a lattice typer.

Reviewed By: wtaha

Differential Revision: D27637721

fbshipit-source-id: f80aa4b860056d4b801e4fdab9dd0b5b46d1ac9c