Start turning on new lattice typer
Summary:
In this diff I start to use the new lattice typer instead of calling `inf_type` on graph nodes. The only place we are using it is in the DOT visualizer used by the fix_problems tests.

In the DOT output the lattice type is the type after the `>=` in the node label.  Notice that in all cases the new lattice typer agrees with the old `inf_type` *except* when the node in question is not supported in BMG. In those cases we now correctly report that the node is "U" for "untypable".  We no longer make any attempt to assign types to nodes that are not even legal in BMG.

Reviewed By: wtaha

Differential Revision: D27527210

fbshipit-source-id: 7f842c9d95a19dee1094f193be5b0b9c3f3b8568