Extract edge requirements to own module (#798)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/798

I am continuing the effort to encapsulate concerns such as node and edge typing into their own modules, rather than spread out amongst all the node classes. This time I'm putting all the edge requirement computations into its own module.  Given a type assignment to each node in a graph, we then can compute requirements on each edge; if the requirements can be met through graph transformations

This is not just a code-reorganizing refactoring; the logic changes as follows:

* Previously we attempted to compute requirements even on input edges of nodes that are unsupported in BMG; we no longer attempt to make guesses about irrelevant requirements. An unsupported node in a graph will either be removed or will cause an error to be reported. If removed, its requirements are irrelevant.  If an error is reported then requirements are never checked at all. So either way it is pointless to write code that computes these requirements.  Unsupported nodes put an "any" requirement on all input edges.

* Same for descendants of unsupported nodes; we make no attempt to compute requirements for all edges that descend from unsupported nodes, for the same reason.

* Computation of requirements often needs to know the type of the node and its input nodes. We now compute that using LatticeTyper rather than calling inf_type on the node.

* Previously we attempted to compute requirements based on substitutions that we knew would be made in the graph in the future. For instance, a power node requires that the exponent input be positive real, or real. But if we knew that the exponent of the power node could be a boolean then we would reason "this power node could be rewritten as IF(exponent, base, 1)" and put a "boolean" requirement on the exponent.  This is silly. The IF node is *a different node*, so why would we say that the *power* requires a boolean here? the *IF* requires a boolean.  We do not make guesses about what transformations will be made in the future; rather, we say what requirements are needed by *the graph that is given* in order to make it compliant with the BMG type system. We don't *need* to know what the requirements are on some hypothetical future edge; we need to know what the requirements are of the *given* edge.

I have not yet updated the requirements checker to use this new logic; rather, I've updated the DOT visualizer to use it, to verify that the changes to the output are the expected ones.

Upcoming work:

In an upcoming diff I will update fix_requirements to use the new type analyzer and requirements computer to assign types and requirements to nodes and edges.

After that I will update (or delete) existing requirements tests, and then delete the requirement and inf_type computation code from the node classes.

I also want to move the requirements logic in bmg_types.py out of there and into bmg_requirements.py.

Finally, the graph_type computation may be unnecessary; it can probably be replaced by a final verification pass and all that code can be deleted.

Reviewed By: wtaha

Differential Revision: D27667711

fbshipit-source-id: a50130823100112dda524446bcaab846dbb26257