Eliminate some unnecessary type analysis (#799)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/799

I want to eventually switch over all uses of inf_type property to use a lattice typer. We can start by eliminating a bunch of unnecessary inf_type calls. A lot of the time we check the type of a thing, we are actually asking "is this the constant zero?" or "is this the constant one?" and in those cases all we have to do is type the value; we don't need a node typer.

I've changed a bunch of the calls to inf_type with helper methods that check to see if we have a constant one or constant zero.

Reviewed By: wtaha

Differential Revision: D27756204

fbshipit-source-id: 6d967af9e18dc99f38dd9256af1deb80a28c41a2