Move complement detection code out of node, into complement rewriting pass
Summary: A graph containing ADD(1, NEG(P)) or ADD(NEG(P), 1) nodes where P is of type Probability or Boolean can be transformed to COMPLEMENT(P).  We originally had the test for whether that transformation could be made in the addition node, but it properly belongs in the rewriter that performs that transformation; whether that transformation is possible or not is not a concern of the node.

Reviewed By: wtaha

Differential Revision: D27756407

fbshipit-source-id: 145f30471d9313b7256e504028ae192703c77538