Integrate typers into problem fixers (#802)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/802

Since I will be removing the ability for nodes to type themselves, the problem fixers will need to be able to obtain the type of a node using a typer.

In a perfect world, each fixer would create the typer it needs, use it to fix the problem, and discard it. However, that introduces a potential performance problem: typing a node can be O(n) in the size of the graph because the type is a function of the types of all the ancestor nodes.

We therefore cache the computed type information in the typer, but that then introduces a second problem: the typer must be informed of when a node is updated, so that if a node type changes, that change is propagated to its descendants.

What we do then is: we create a single typer at the start of the problem fixing pass and give it to each of the fixers. A fixer is required to update the typer when it makes a change to the graph.

This scheme introduces some possibilities for bugs -- a fixer might forget to make an update, which causes a wrong type to be deduced later, for instance.  To mitigate this possibility I've put the updating logic in the fixer base class so that we have only one place where we need to get it right.

The requirements fixer does not inherit from that base class yet -- I mean to refactor it later.  In the meanwhile I have ensured that it does update the typer appropriately.

None of the fixers *use* the typer yet. This diff just gets the underlying plumbing working.  In the next diffs I will change the problem fixers over to use the typers rather than calling inf_type on the nodes.

Reviewed By: wtaha

Differential Revision: D27757065

fbshipit-source-id: 86fe590c25ca3f9fbd95678be12a1f3e77daf53e