Arithmetic rewriters now use lattice typer
Summary:
The arithmetic rewriters now use the lattice typer to detect rewritable sums, products, powers and so on.

Since "is this usable as a bool? probability?" are common questions, I've added a helper method for those to the typer, thereby making the code read more clearly at the call site.

Reviewed By: wtaha

Differential Revision: D27779492

fbshipit-source-id: 63696479aff9e725c5e6bfc010b7902c98897bd3