Arithmetic rewriters now use lattice typer (#803)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/803

The arithmetic rewriters now use the lattice typer to detect rewritable sums, products, powers and so on.

Since "is this usable as a bool? probability?" are common questions, I've added a helper method for those to the typer, thereby making the code read more clearly at the call site.

Reviewed By: wtaha

Differential Revision: D27779492

fbshipit-source-id: 38fe6f9a55e80724e2c8b096f376f511c5981ed4