Requirements fixer now uses lattice typer and edge requirements objects (#805)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/805

The requirements fixer now uses a lattice typer and and edge requirements analyzer to figure out how to insert conversion nodes.

This is almost the last use of the inf_type property, and the last use of the requirements property on nodes outside of test cases. In upcoming diffs I will remove the last few uses of inf_type and delete all the now-unused code.

Reviewed By: wtaha

Differential Revision: D27782902

fbshipit-source-id: becb4f6f8ac124fada244ff8c932579c882d71f6