Eliminate all callers of graph_type property
Summary:
The graph_type property on nodes is now called nowhere; I will remove it in an upcoming diff.

In a small number of cases we need to distinguish between the lattice type of a constant -- how small a type can this value fit into? -- and the type that BMG associates with an actual node. I've put the logic for those cases in a method `graph_type` of the lattice typer.

There are a couple of small changes in the value computed by `graph_type`:

* I've eliminated generation of the Malformed type annotation, which was used for situations such as an addition where the addends were not of the same type. We do not actually need to track this in the type system, and I will remove it in an upcomding diff as well.  Nodes which were previously marked as malformed are now considered to have their lattice type for display purposes.

* Nodes which are classified as untypable by the lattice typer are similarl classifed as untypable by `graph_type`.

Reviewed By: wtaha

Differential Revision: D27806244

fbshipit-source-id: 3b7d869c2ea965f6349b6f095214d82b6ba3bc72