Start eliminating graph_type altogether from compiler
Summary:
In this diff and the next I will eliminate the graph_type feature from the compiler entirely, and simplify how constant nodes are created and managed.

Let's start by describing the existing design and its shortcomings.

In a Python model, almost all relevant non-stochastic values are float-valued single-valued tensors. There is no particular type system imposed upon them.

In Bean Machine Graph by contrast, there are many different nodes that represent constant values: we have distinct nodes for bool, natural, negative real, positive real and real values, and similarly for two-dimensional fixed-size matrices of those types.

Moreover, the actual values used in a Python model are different; a Bernoulli coin flip in a Python model for example will produce a tensor(1.0) value for "heads", but in BMG that thing is the Boolean "true".

This presents a problem to be solved when transforming a Python model to BMG; when we encounter a constant value that is a parameter to a stochastic computation, how do we know what kind of node to generate?

The original strategy was as follows:

* Transform the Python program into a form that, when executed, calls back graph accumulation methods on every operation.

* Suppose for example we have an addition of 1.5 to a draw from a half Cauchy. We create nodes for the distribution, for the sample, for the addition, and for the constant 1.5. We generate a real-valued node for the constant.

* When it comes time to check the type of each node, we note that the addition is adding a positive real sample from a half Cauchy to a real constant 1.5, but that constant value is compatible with the positive real type. We replace the real constant with a positive real constant.

This means that for every constant node we need to effectively keep track of two things: what is the type of the *node*, from BMG's perspective, and what is the type of the *value*, so that we know what kinds of constant nodes we could convert this value to if so required.  I referred to the former as the "graph type" and the latter as the "inf type", and we had two mechanisms for computing each when we needed them.

This is needlessly complicated; what is the source of the complication? **That a real-valued constant node can be treated as having a smaller type sometimes, depending on its value**.

But there is another problem here too. As much as possible I wish to keep the *type semantics of BMG* separate from *decisions made by the graph accumulator at accumulation time*.  The BMG type system should be relevant when *deciding how to convert an accumulated graph into a BMG graph*, and not relevant when simply executing a model and recording what it does.

I therefore have removed creation of "typed" BMG nodes completely from the graph accumulation step.  When the model is executing and we are accumulating nodes, we always accumulate UntypedConstantNode nodes.  We regularize the value to bool, int, float or tensor, we deduplicate based on both value and type, but we do NOT attempt to assign a BMG node type to the untyped constant node.

When we assign types to each node during the problem fixing phase we associate the "inf type" -- the type associated with the *value* -- to untyped constant nodes. When we compute an edge requirement on an edge where an input is an untyped constant, we can compare the requirement to the type of the value; if the requirement can be met then we can generate a typed node containing that value.

In this system there is no difference between the "graph type" and "inf type" of any node, so for now I have assigned them to be equal. You can see this reflected in the DOT test output which shows both graph and inf types:

* Graph and inf types are now always identical
* Type of an "untyped" const node is the type of its value
* Type of a "typed" const node is the type of the node

You can also see by looking at the graph output:

* We are now "orphaning" all the untyped constant nodes. Every untyped constant node which outputs to a supported graph node is replaced with the equivalent typed node, leaving the untyped constant as the ancestor of no sample, query or observation.

In the next diff I will delete the graph_type feature and redo every test case which displays this now-redundant information.

Reviewed By: wtaha

Differential Revision: D27921722

fbshipit-source-id: 8c4a1b2f78e9031b8b98a7abdcc6d127d7768a00