Refactor worlds (#810)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/810

This diff does nothing special by its own other than splitting the current `World` class into a `BaseWorld` and a single-site specific `World`. A few utility functions are also refactor into `utils.py` so that they can be used by other "worlds".

The change in this diff enable us to do different things with the model by subclassing from the `BaseWorld`. When a `bm.random_variable` function is invoked, it will look up the world instance from the closest context and invoke the `update_graph()` method.

In this diff the `BaseWorld` class only holds the the context manager interface because I can't find a good way to utilize the other methods that are currently defined in single site `World` and adapt them to the global world (which will be introduce in the next diff in the stack, D28215033).

So in a way the purpose of this diff is just to open up an API that allow me to redirect the random variable decorator so I can start developing a relatively independent global inference framework without changing a ton of existing codebase.

Reviewed By: neerajprad

Differential Revision: D28215010

fbshipit-source-id: adc0c7c5b2c234ee60db85611cd0c06f53ea88c7