Completely eliminate graph type feature from compiler (#814)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/814

This finishes off the work of reforming the overall typing system of the compiler. As noted in the previous diff, we used to keep track of "graph type" and "inf type" separately due to an oddity in the way we constructed constant nodes. Now that this oddity has been eliminated, we can unify these two related ideas and eliminate one of them.

Reviewed By: wtaha

Differential Revision: D27926797

fbshipit-source-id: 2a9fecd2b3d96097ec0f48dafd819122c229d705