Add support for TO_MATRIX operation in compiler (#815)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/815

Lily recently added a TO_MATRIX operator in BMG which takes some number of graph nodes and forms a matrix from them. We will need to generate this operator in the compiler so I've added support for it.

So far all we have demonstrated is that we can add such a node to the graph builder and then:

* Label the node and edges correctly on a DOT output
* Compute the type of the node and the requirements on its input edges
* Generate correct C++ and Python code to make the corresponding BMG graph
* Generate the corresponding BMG graph object without error, and convert it to DOT as well.

We already accumulate TensorNode objects which appear to be exactly the same thing, but they are not quite. For example, since BMG is column major but pytorch is row-major, we will need to take the transpose when we generate the TO_MATRIX node from a tensor node.

I have not yet added support for Lily's indexing operation on multidimensional tensor, nor for the tensorized version of logsumexp; all that will come in future diffs.

Reviewed By: wtaha

Differential Revision: D27829449

fbshipit-source-id: f2471c33fe494b192f1eab54e2cf6f1ac7e45a9d