Start adding support for LOGSUMEXP_VECTOR BMG node in compiler (#817)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/817

Lily has recently added a version of LOGSUMEXP to BMG which takes a single vector as its input rather than n individual inputs.  This is useful in scenario such asL a model has a 2d vector, we choose a 1d column vector out of it using COLUMN_INDEX, and then we log-sum-exp that column.

There will be several steps needed to get support for this operation in the compiler and I'm going to do it in several diffs.

In this first diff, I simply add an equivalent node type to the compiler, along with support for determining its node type, its input edge requirements, and so on.

Looking ahead, we will require additional work to fully support this operation.  Likely what I will do is:

* Create a *third* logsumexp node that specifically represents "we have a call to tensor.logsumexp(dim) in the original python model"
* Update the error detection pass to detect models where the user has passed in a stochastic value for the dim argument; we only support models where we have a constant zero for this argument.
* Update the rewriting pass to decide whether a logsumexp(dim) operation should be converted to the vectorized or unvectorized version; obviously right now we only support the unvectorized version but that will need to change.

Reviewed By: wtaha

Differential Revision: D27835737

fbshipit-source-id: a01084aa92db478bc6fd874c02d4fda7f6b0a465