Start adding support for COLUMN_INDEX in compiler
Summary:
Lily has recently added a COLUMN_INDEX operator to BMG which takes as its input a 2d matrix and an integer, and produces a column matrix.

We will eventually need to support generating this from the compiler; the first step is adding a node type and getting its type analysis correct.  Here we demonstrate that we can construct a 2x2 matrix with the TO_MATRIX operator, and then randomly choose one of its columns based on a coin flip to get a 2x1 column vector, which can then be passed to the LOGSUMEXP_VECTOR operator added in the previous diff.

{F611018262}

Reviewed By: wtaha

Differential Revision: D27870459

fbshipit-source-id: 830ec7a585f1bcc4a10689fdc5a76d46e456bd41