Fix row/column mismatch in type system (#819)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/819

Torch, and therefore Bean Machine, uses tensors in row-major order; for instance, if we are setting the parameters of a Dirichlet distribution to 1, 2, 3 then we expect the tensor to be `[1, 2, 3]`.  One row, three columns.

Bean Machine Graph by contrast prefers column major order; the input to a Dirichlet must be `[[1], [2], [3]]`  One column, three rows.

The BMG type system has scalar types and 2-d matrix types. How then should we assosciate matrix types with nodes?  Should we use the type of the *storage at runtime* and say that this vector of dirichlet parameters is a 1x3 matrix because it is a 1x3 tensor?  Or should we use the type of the *storage in the emitted graph* and call it a 3x1 matrix?

We were originally doing the former, but this becomes quite confusing when we need to check for requirements such as "the input to a Dirichlet must have exactly one column".

I have modified everywhere in the type system (I hope!) that determines the type of a matrix value; it now states the row x column size of the matrix as it will be in the BMG type system. After all, that is what we are analyzing here: what will the types of these nodes be in that type system?

Graph accumulator nodes still have a size property that I need to refactor away; this gives the size of the node using the conventions of the Bean Machine runtime.  But the computation of this property is (1) recursive, which is bad, and (2) frequently wrong. I will likely do a refactoring later to move this computation into its own type analyzer.

A question we will need to answer when we design a proper error reporting system is whether to phrase the errors in the jargon of the BMG type system or the BM type system. Right now we use the BMG type system but we might want to revisit that decision.

Reviewed By: wtaha

Differential Revision: D28066184

fbshipit-source-id: 9dd2f4d0e194f9f3419208bffd67aa4d58bf9ca0