Implement TO_REAL_MATRIX and TO_POS_REAL_MATRIX in BMG (#820)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/820

In BMG we had no way to easily turn a matrix of, say, probabilities, into a matrix of reals should the type system require one.

I've added operators TO_REAL_MATRIX and TO_POS_REAL_MATRIX which take a matrix as an input and produce the equivalent real / pos real matrix respectively as the output.

There are two more we might need for rare cases: convert a matrix of bools to a matrix of naturals, and convert a matrix of bools to a matrix of probabilities. We haven't run into a situation where we need them yet.

I have not yet implemented the code in the compiler to take advantage of these new operators; that will happen in the next diffs.

Reviewed By: yucenli

Differential Revision: D28081087

fbshipit-source-id: 12480461cb37574883488b8db88bee9c44e1d8c3