Refactor is_matrix operation into type checker (#821)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/821

We need to keep track of whether a particular node is typed as being an atomic value, say, real, or as a 1x1 matrix of reals.  (I am somewhat regretting my early decision to make the atomic types and the 1x1 matrix types represented by the same object; this is not how BMG does it, and it might be smart to make the two type systems more similar at the implementation choices level.  I'm not going to address that design problem here though.)

That fact was represented with a flag on the node, but it properly should be an API on the type analyzer. It now is.

Reviewed By: wtaha

Differential Revision: D28128342

fbshipit-source-id: 2fcb0ba1f73b9bba08eca7a2874bf83391d3de5a