Small refactoring in requirement fixer (#823)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/823

I'm going to make some changes to support the new TO_REAL_MATRIX and TO_POS_REAL_MATRIX operators in the requirements fixer, so now is a good time to make a few small improvements to clarify the code.

The basic change here is to move the check for "is the requirement already met?" earlier, and to turn some of the checks for conditions that should be impossible into assertions.

In an upcoming diff I'll reorganize `meet_operator_requirement` to be more clear and to support the new conversion operators.

Reviewed By: wtaha

Differential Revision: D28165808

fbshipit-source-id: e23ef4731615f4ca84557288c4353b6566e8704b