Fix bug in tensor constructor handling (#827)
Summary:
Pull Request resolved: https://github.com/facebookincubator/beanmachine/pull/827

The tensor constructor is allowed to take a single ordinary value, such as `tensor(123)`, in which case its size is `Size([])`, or it can take a list of values, or a list of lists of values, and so on.

There was a bug in how I was handling the tensor constructor where I always treated it as one level more nested than it should have been. That is, I allowed `tensor(1, 2, 3)` even though that should be illegal, because I treated it like `tensor([1, 2, 3])`.  Similarly I treated `tensor([1, 2, 3])` as `tensor([[1, 2, 3]])`, and so on.

This is now fixed; we now verify that the tensor constructor has exactly one argument, and compute the size correctly.

Reviewed By: wtaha

Differential Revision: D28334386

fbshipit-source-id: fa7eeb0fa54591980f363fd17bd8b7087e97c2fe