Compiler now generates appropriate index operation
Summary:
When a model contains an `[]` operator we now accumulate it into the graph as an `IndexNode` that is NOT supported by BMG.

In the problem fixing pass we determine if it is an index into a 1-column vector or a multi-column vector, and replace the  `IndexNode` with `VectorIndexNode` or `ColumnIndexNode` as appropriate.

Optimizations for scenarios where the index is a constant are moved into the fixer rather than the factory. Generally speaking I'd like to move these peephole optimizations out of the factories; the factories should not need to know the semantics of the operators.

Reviewed By: wtaha

Differential Revision: D28364431

fbshipit-source-id: 459000987d4c79c41bbe0d02956fc89cf4fd9762