Add LSE_VECTOR support to the compiler
Summary:
The compiler now generates LOGSUMEXP operations in two ways.

First, it can generate an n-ary LOGSUMEXP operator when its operand is a TO_MATRIX containing a single column.  We discard the TO_MATRIX and input the elements directly into the LSE.

Second, if we have an LSE operation where the operand is a single column vector but is NOT an TO_MATRIX then we generate a LOGSUMEXP_VECTOR node.

We previously had code to do the former but not the latter. I've reworked the code so that both rewritings are done in the "eliminate unsupported nodes" fixer.  This meant that the tensor operations fixer was now doing nothing at all, so I've deleted it.  (At some point we will probably want to do optimizations or rewrites on tensor ops; when that happens I will resurrect the code. But there is no need to keep useless code around in case we might need it later.)

Differential Revision: D28372150

fbshipit-source-id: c32ef289fb68005ca4b8cc9a687435a1403a6857