Extending tests to be better localize problem in single_site_ancestral_mh_conjugate_test_nightly.py
Summary: This diff introduces a set of ten seeds to run the various tests on. The distant_normal case fails on all ten values. This further supports the idea that the issue is with distant_normal_normal case and not necessarily the algorithm itself. In all cases it fails with the error indicated in the comment on the test (low ESS, and exactly the same value on both seeds). This suggests that the algorithm is consistently getting stuck on this case.

Differential Revision: D28519753

fbshipit-source-id: ece181f6aa8334fe7c193b9934dbb5bd57b3d6c9