Extending tests to be better localize problem in single_site_ancestral_mh_conjugate_test_nightly.py
Summary: This diff introduces a set of ten seeds to run the various tests on. This further supports the idea that the issue is with distant_normal_normal case and not necessarily the algorithm itself. That test consistently fails on all ten randomly selected random_seeds. The specification of the test is now that the failure is reproducible on all ten random_seed values.

Differential Revision: D28519753

fbshipit-source-id: eb1d08551b429a3812edc3ed6bcd6a061f3a22ac